Add teacher service

+++ auth-service/controllers/authController.js
@@ -4,7 +4,7 @@ import bcrypt from 'bcrypt'
 import TokenRepository from "../repositories/tokenRepository.js";
 import UserRepository from '../repositories/userRepository.js';
 import { asyncWrapper } from "#shared/middlewares/index.js"
-import { BadRequestError, ConflictError, NotFoundError, UnauthorizedError, InternalServerError } from "#shared/errors/errors.js";
+import { BadRequestError, ConflictError, NotFoundError, UnauthorizedError, InternalServerError, ForbiddenError } from "#shared/errors/errors.js";
 import { JwtFacade } from '../utils/jwt.js';
 import createHash from '../utils/createHash.js';
 
@@ -20,20 +20,17 @@ const createNewRefreshToken = async (token, userId) => {
 
 export const token = asyncWrapper(async (req, res) => {
   const tokenCookie = req.cookies.refreshToken;
-  if (!tokenCookie)
-    return res.status(404).json({ msg: "No token in cookie!" });
+  if (!tokenCookie) throw new NotFoundError("No token in cookie!");
 
   const isRefreshToken = await TokenRepository.findOne({ refreshToken: tokenCookie })
-  if (!isRefreshToken)
-    return res.status(404).json({ msg: "Token is not available!" });
+  if (!isRefreshToken) throw new ForbiddenError("Invalid or expired token!");
 
   let user;
   try {
     user = JwtFacade.verifyToken(tokenCookie, process.env.REFRESH_TOKEN_SECRET);
-    console.log("User: ", user);
   } catch (err) {
     await TokenRepository.deleteByToken(tokenCookie); // Xóa token trong db nếu expired
-    return res.status(401).json({ msg: "Invalid or expired refresh token" });
+    return res.status(401).json({ msg: "Invalid or expired token!" });
   }
 
   // Rotate refreshToken
@@ -47,9 +44,7 @@ export const token = asyncWrapper(async (req, res) => {
 export const register = asyncWrapper(async (req, res) => {
   const { username, password } = req.body;
 
-  if (!username || !password) {
-    throw new BadRequestError("Missing required field") // 400
-  }
+  if (!username || !password) throw new BadRequestError("Missing required field!") // 400
 
   const user = await UserRepository.findOne({ username: username })
 
@@ -69,16 +64,13 @@ export const register = asyncWrapper(async (req, res) => {
 export const login = asyncWrapper(async (req, res) => {
   const { username, password } = req.body;
 
-  if (!username || !password)
-    throw new BadRequestError("Missing requires field");
+  if (!username || !password) throw new BadRequestError("Missing requires field!");
 
   const user = await UserRepository.findOne({ username: username });
-  if (!user)
-    throw new NotFoundError("Username not found!");
+  if (!user) throw new NotFoundError("Username not found!");
 
   const isPassword = await bcrypt.compare(password, user.password);
-  if (!isPassword)
-    throw new UnauthorizedError("Invalid password!");
+  if (!isPassword) throw new UnauthorizedError("Invalid password!");
 
   const { accessToken, refreshToken } = await JwtFacade.setTokens(res, user._id);
   await createNewRefreshToken(refreshToken, user._id)
@@ -90,11 +82,8 @@ export const changePassword = asyncWrapper(async (req, res) => {
   const { password } = req.body;
   const userId = req.user.userId;
 
-  if (!userId)
-    throw new InternalServerError("Required userId is invalid!")
-
-  if (!password)
-    throw new BadRequestError("Missing requires field");
+  if (!userId) throw new InternalServerError("Required userId is invalid!");
+  if (!password) throw new BadRequestError("Missing requires field!");
 
   const hashedPassword = await createHash(password);
   await UserRepository.updatePasswordById(userId, hashedPassword);
@@ -104,10 +93,7 @@ export const changePassword = asyncWrapper(async (req, res) => {
 
 export const logout = asyncWrapper(async (req, res) => {
   const refreshToken = req.cookies.refreshToken;
-  if (!refreshToken)
-    return res.status(404).json({ msg: "No token in cookie" });
-
-  console.log(refreshToken);
+  if (!refreshToken) throw new NotFoundError("No token in cookie!");
 
   res.clearCookie("refreshToken", {
     httpOnly: true,
@@ -115,9 +101,8 @@ export const logout = asyncWrapper(async (req, res) => {
     sameSite: "strict",
   });
 
-  const isRefreshToken = await TokenRepository.findOne({ refreshToken: refreshToken })
-  if (!isRefreshToken)
-    return res.status(403).json({ msg: "Invalid or expired token" });
+  const isRefreshToken = await TokenRepository.findOne({ refreshToken: refreshToken });
+  if (!isRefreshToken) throw new ForbiddenError("Invalid or expired token!");
 
   await TokenRepository.deleteByToken(refreshToken);
 
@@ -127,11 +112,7 @@ export const logout = asyncWrapper(async (req, res) => {
 export const getUsernameByIds = asyncWrapper(async (req, res) => {
   const ids = req.query.ids ? req.query.ids.split(",").filter(id => id.trim() !== "") : [];
 
-  if (!ids || ids.length === 0)
-    return res.status(400).json({
-      success: false,
-      msg: "Missing ids"
-    });
+  if (!ids || ids.length === 0) throw new BadRequestError("Missing user id!");
 
   const users = await UserRepository.findNameByIds(ids);
